fixed warnings

+++ src/nuspell/dictionary.hxx
@@ -56,7 +56,7 @@ class Dictionary {
 	auto spell_casing_title(std::basic_string<CharT>& s) -> const Flag_Set*;
 	template <class CharT>
 	auto spell_sharps(std::basic_string<CharT>& base, size_t n_pos = 0,
-	                  int n = 0, int rep = 0) -> const Flag_Set*;
+	                  size_t n = 0, size_t rep = 0) -> const Flag_Set*;
 	template <class CharT>
 	auto checkword(std::basic_string<CharT>& s) -> const Flag_Set*;
